requirements.txt
ollama~=0.4.4
discord~=2.3.2
python-dotenv~=1.0.1